pyproject.toml
[project]
name = "final-tcc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "deeplabcut[gui,modelzoo]>=3.0.0rc13",
    "faiss-cpu>=1.7.2",
    "opencv-python>=4.8.0",
    "pyside6>=6.10.0",
    "pyyaml>=6.0.3",
    "tables>=3.9.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]

[dependency-groups]
dev = [
    "hypothesis>=6.100.0",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-qt>=4.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["src"]
omit = ["src/ui/tabs/*", "src/ui/main_window.py"]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]
